Retire the common-mode mechanism against its own null and its trained operator
The second hop's failure was attributed to common-mode swamping: `a @ a` was
measured carrying 0.7550 of its energy in a rank-1 mode against 0.4899 for the
K=8 routed form, and the routing was read as an accidental filter. Both figures
were taken on an untrained operator and neither was compared to a null. Against
random Gaussian logits at the measured logit sd of 0.0375, causal-masked and
softmaxed over 20 seeds, cm_full(a@a) has a null of 0.7550 +/- 0.0004 against a
measured 0.7549, z = -0.30. Training one arm 150 steps moves the logit sd from
0.0367 to 2.7645, a factor of 75.4, and the readout-row common-mode fraction
from 0.8328 to 0.3736. The mechanism is withdrawn in full.

What the withdrawal covers
The depth-compounding table, the task-invariance table, the K=8 versus K=64
common-mode comparison at both full-matrix and readout-row resolution, and the
deflation-at-depth table. Each remains a correct statement about an untrained
causal-softmax operator at s=64 and none describes the arms after training. The
task-invariance result is the clearest case: cm(a@a) agreeing to four figures
across e3_t2, e3_t8 and e3_t32 was recorded as evidence of construction-level
structure, and a quantity constant across every condition varied and equal to
its own null is the baseline.

What the withdrawal does not cover
Every trained reading stands. K sweep at t*=2, n=2048, 3 seeds, threads=8:
softmax 0.950252, K=8 0.960945, K=16 0.968518, K=32 0.965750, K=64 1.000329.
Gain sweep at K=64, same cell: gain 0.00 reproduces the softmax control at
+0.000000, an exact identity check, then +0.008261, +0.006960 and +0.002384 at
gain 0.05, 0.10 and 0.25, every contrast inside one control sd of 0.015381.
`max|softmax - pivot_unsigned| = 0.000e+00` is structural and weight-independent.

Verdict machinery
`scale/it11_verdict.py` gains a short-N refusal on both TOST samples: fewer
seeds widen the interval, and a wide interval inside a wide margin still reads
PARITY, so a short N makes an equivalence test easier to pass.
`scripts/v13_adjudicate_hop2.py` reads the four clauses of the filed hop-2
prediction off the journal and refuses below N=8; it was written while the cell
stood at 4 of 8 seeds, before the outcome existed.

Independent audit
`V13_CLAIM_AUDIT.md` re-derived 54 claims: 44 CONFIRMED, 9 DISCREPANT, 1
UNVERIFIABLE. N=70 was confirmed by three routes sharing no code. All nine
discrepancies are corrected here, including a withdrawn harness control that
crossed thread lanes and whose true gap of 0.002097 is 0.894 of the 2.345e-3
reduction-order floor. B9's throughput figures are reopened: they came from a
3-step run rather than the specified 500, and `results/r10_v13_b9_4060_probe.txt`
is 0 bytes, so no artifact survives. `scripts/v13_b9_4060_probe.py` now passes
dtype="fp32" to match the loop it measures.

Prior art
Six fetch reports land. `κ = λ(1 − d)` uses the information dimension D1, not
box-counting D0, and D1 <= D0 always, so the substitution under-predicts the
escape rate by 0.49% to 24.5% across slope ratios 1.5 to 500; the contract's
worked figure sits on the constant-slope line where the two coincide and the
defect cannot appear. DC-mode attenuation is standard graph signal processing,
orthogonal-code multiplexing is the VSA/HRR construction, and published
crosstalk scaling is sqrt(M/L) rather than 1/sqrt(L).

Limits
MISTAKES.md now carries 53 entries and workdonenew.md section 5 carries 32 rows;
9 entries and 18 rows are from this session. The K and gain sweeps run 3 seeds,
below the round's registered N=8, and no verdict is claimed from either. The
t*=8 n=32768 deciding cell stands at 5 of 8 seeds and its filed prediction is
unadjudicated. `V13_X29A_DEFLATION.md` section 8 is still SWEEP_TABLE_PLACEHOLDER,
so the deflation recovery number does not exist. Round 11's reading has not run,
and no C-PAR or C-CAP verdict is claimed.

+++ scripts/v13_b9_4060_probe.py
@@ -94,7 +94,14 @@ def main() -> int:
                 peak, rate = peak_and_rate(kind, d=a.d, layers=layers, heads=a.heads,
                                            seq=a.seq, bs=a.bs, device=dev,
                                            steps=a.steps, ckpt=ckpt)
-                pred = sz.activation_bytes(cfg, batch=a.bs, arm=kind, checkpointed=ckpt)
+                #: The loop below runs in fp32 -- there is no autocast anywhere in
+                #: `peak_and_rate` -- so the prediction must be read at fp32 too.
+                #: An earlier version left `sizing.activation_bytes` at its
+                #: `bf16_autocast` default and printed a prediction column that
+                #: disagreed with its own measurement by the dtype ratio, which
+                #: made M-12's fp32 figures look wrong when reproduced from here.
+                pred = sz.activation_bytes(cfg, batch=a.bs, arm=kind,
+                                           dtype="fp32", checkpointed=ckpt)
                 if peak is None:
                     print("%-8s %-3d %-6s %12s %12.1f %8s %12s"
                           % (kind, layers, ckpt, "OOM", pred / 2**20, "-", "-"))
